Add list routes
- Getting the List index only returns lists that the user owns or is subscribed to
- A user can only update or delete a list that they own
- Added testing around that principle

NOTE:

For some reason (probably a bug in the JSON:API resource gem being used), the validation of ownership when adding or deleteing a List cannot be done from within the resource itself.  Instead, this logic must live inside the controller.  While this isn't a bad thing, it would be nice to move this logic into the Resource once that is made possible.