Fix resource URL naming
As it turns out, I messed up the naming of the User and List
controllers, which made the URLS be singular instead of plural.  This
issue reared it's head one I tried to test that you were unable to
update or delete a list's owner; the URLs for those actions were
generated with the plural resource names, not the singular ones.

This batch of changes re-names the controllers and fixes the routing, as
well as adding a catch-all route for any */api* URLs or actions that are
not support by the application, rendering a 404 error in those cases.